Move tinabot to pkg directory.
