+++ pkg/tinabot/tinabot_test.go
