+++ pkg/tinabot/userchoice.go
