+++ pkg/tinabot/userchoice_test.go
